Experiment with Interactrion Network, change data input format

+++ Code/Utilities/M2b_TrainModel_Sub.py
@@ -138,7 +138,7 @@ def train(Predict, model, device, sample, optimizer, epoch):
 
 
 
-        y, w = data.y, w.squeeze(1)
+        y, w = data.y.float(), w.squeeze(1)
         print(y)
         print(w)
     #     particle_id = data.particle_id
